# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorCode": {
        "description": "The stable wire-level failure codes currently emitted by the control plane.",
        "enum": [
          "ADMINISTRATION_WOULD_BE_LOST",
          "LOGIN_NAME_TAKEN",
          "PASSWORD_TOO_SHORT",
          "PERMISSION_UNREGISTERED",
          "ROLE_CODE_TAKEN",
          "ROLE_NOT_FOUND",
          "USER_NOT_FOUND",
          "ACCOUNT_DEACTIVATED",
          "AUTHENTICATION_REQUIRED",
          "CREDENTIALS_REJECTED",
          "CSRF_TOKEN_INVALID",
          "INFERENCE_BACKEND_ENDPOINT_TAKEN",
          "INFERENCE_BACKEND_NOT_FOUND",
          "INFERENCE_HOST_DEACTIVATED",
          "INFERENCE_HOST_HAS_BACKENDS",
          "INFERENCE_HOST_NAME_TAKEN",
          "INFERENCE_HOST_NOT_FOUND",
          "INTERNAL_ERROR",
          "PERMISSION_DENIED",
          "REQUEST_INVALID",
          "SESSION_INVALID",
          "STALE_REVISION"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
...
      "BackendPlacement": {
        "description": "The complete placement of one process endpoint.",
        "properties": {
          "base_url": {
            "maxLength": 255,
            "minLength": 1,
            "pattern": "^https?://",
            "title": "Base Url",
            "type": "string"
          },
          "host_id": {
            "format": "uuid",
            "title": "Host Id",
            "type": "string"
          }
        },
        "required": [
          "host_id",
          "base_url"
        ],
        "title": "BackendPlacement",
        "type": "object"
      },
      "BackendStatus": {
        "description": "The active/deactivated value of the backend status subresource.",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "BackendStatus",
        "type": "object"
      },
...
      "InferenceBackendView": {
        "description": "One process endpoint and its single reserved template-binding slot.",
        "properties": {
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "connection": {
            "$ref": "#/components/schemas/ConnectionView"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "host_id": {
            "format": "uuid",
            "title": "Host Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          },
          "template_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Version Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "format": "uuid",
            "title": "Updated By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "host_id",
          "base_url",
          "template_version_id",
          "status",
          "connection",
          "revision",
          "created_by",
          "updated_by",
          "created_at",
          "updated_at"
        ],
        "title": "InferenceBackendView",
        "type": "object"
      },
...
      "ItemPage_InferenceBackendView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InferenceBackendView"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "ItemPage[InferenceBackendView]",
        "type": "object"
      },
...
      "ProblemDocument": {
        "description": "RFC 9457 plus the stable control-plane extensions from §5.15.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error_code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "about:blank",
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "error_code"
        ],
        "title": "ProblemDocument",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SOP compliance center backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inference-backends": {
      "get": {
        "description": "List endpoints, optionally filtered to one host, under the common page envelope.",
        "operationId": "listInferenceBackends",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "host_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Host Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemPage_InferenceBackendView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List The Backends",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_backend.view"
      },
      "post": {
        "description": "Register an endpoint on an existing active host.",
        "operationId": "createInferenceBackend",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackendPlacement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceBackendView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Host not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Host deactivated, or the endpoint is taken"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create A Backend",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_backend.edit"
      }
    },
    "/api/v1/inference-backends/{backend_id}": {
      "delete": {
        "description": "Delete one endpoint outright; deactivation is the reversible alternative.",
        "operationId": "deleteInferenceBackend",
        "parameters": [
          {
            "in": "path",
            "name": "backend_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Backend Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Backend not found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Delete A Backend",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_backend.delete"
      },
      "get": {
        "description": "Read one endpoint, including connection and provenance facts.",
        "operationId": "readInferenceBackend",
        "parameters": [
          {
            "in": "path",
            "name": "backend_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Backend Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceBackendView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Backend not found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Read A Backend",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_backend.view"
      },
      "patch": {
        "description": "Replace placement at the revision the caller read.",
        "operationId": "editInferenceBackend",
        "parameters": [
          {
            "in": "path",
            "name": "backend_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Backend Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackendPlacement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceBackendView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Backend or host not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Revision moved (STALE_REVISION), host deactivated, or endpoint taken"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Edit A Backend",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_backend.edit"
      }
    },
    "/api/v1/inference-backends/{backend_id}/connection-test": {
      "post": {
        "description": "Ask the endpoint once and persist the actual connection result.",
        "operationId": "testInferenceBackendConnection",
        "parameters": [
          {
            "in": "path",
            "name": "backend_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Backend Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceBackendView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Backend not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Revision moved (STALE_REVISION)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Test A Backend Connection",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_backend.edit"
      }
    },
    "/api/v1/inference-backends/{backend_id}/status": {
      "put": {
        "description": "Set the reversible active/deactivated status.",
        "operationId": "setInferenceBackendStatus",
        "parameters": [
          {
            "in": "path",
            "name": "backend_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Backend Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackendStatus"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceBackendView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Authentication required or session invalid"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Permission denied or CSRF token invalid"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Backend not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Revision moved (STALE_REVISION)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Request invalid"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Set The Backend Status",
        "tags": [
          "device"
        ],
        "x-required-permission": "device.inference_backend.edit"
      }
    },
...
  }
}